Fix FLX field masking impl fails with int values outside of the 2³² range.
Closes #154

See merge request search-guard/search-guard-suite-enterprise!326

+++ dlic-dlsfls/src/main/java/com/floragunn/searchguard/enterprise/dlsfls/lucene/FlsStoredFieldVisitor.java
@@ -233,11 +233,11 @@ class FlsStoredFieldVisitor extends StoredFieldVisitor {
                         break;
 
                     case VALUE_NUMBER_FLOAT:
-                        generator.writeNumber(parser.getFloatValue());
+                        generator.writeNumber(parser.getDecimalValue());
                         break;
 
                     case VALUE_NUMBER_INT:
-                        generator.writeNumber(parser.getIntValue());
+                        generator.writeNumber(parser.getBigIntegerValue());
                         break;
 
                     case VALUE_STRING:
